feat: Global chat refactor
